added likes and delete functions

+++ src/utils/api.js
@@ -54,18 +54,32 @@ class Api {
 			headers: this.headers,
 		});
 	}
-	async likeCard(cardId) {
-		return this._request(`${this.url}/cards/likes/${cardId}`, {
-			method: "PUT",
-			headers: this.headers,
-		});
+	// async likeCard(cardId) {
+	// 	return this._request(`${this.url}/cards/likes/${cardId}`, {
+	// 		method: "PUT",
+	// 		headers: this.headers,
+	// 	});
+	// }
+	// async unLikeCard(cardId) {
+	// 	return this._request(`${this.url}/cards/likes/${cardId}`, {
+	// 		method: "DELETE",
+	// 		headers: this.headers,
+	// 	});
+	// }
+	async changeLikeCardStatus(cardId, isLiked) {
+		if (!isLiked) {
+			return this._request(`${this.url}/cards/likes/${cardId}`, {
+				method: "PUT",
+				headers: this.headers,
+			});
+		} else {
+			return this._request(`${this.url}/cards/likes/${cardId}`, {
+				method: "DELETE",
+				headers: this.headers,
+			});
+		}
 	}
-	async unLikeCard(cardId) {
-		return this._request(`${this.url}/cards/likes/${cardId}`, {
-			method: "DELETE",
-			headers: this.headers,
-		});
-	}
+
 	async changeProfilePicture(link) {
 		return this._request(`${this.url}/users/me/avatar`, {
 			method: "PATCH",
